requirements.txt
aiogram==3.10.0
aiohttp==3.9.5
requests==2.31.0
python-dotenv==1.0.0
pytz==2024.1
matplotlib==3.8.2
pandas==2.1.4
numpy==1.24.4
openpyxl==3.1.2